1. Созданы модули :expense и :income
