+++ .idea/gradle.xml
@@ -15,6 +15,8 @@
             <option value="$PROJECT_DIR$/features" />
             <option value="$PROJECT_DIR$/features/cash_account" />
             <option value="$PROJECT_DIR$/features/category" />
+            <option value="$PROJECT_DIR$/features/expense" />
+            <option value="$PROJECT_DIR$/features/income" />
             <option value="$PROJECT_DIR$/features/settings" />
             <option value="$PROJECT_DIR$/navigation" />
             <option value="$PROJECT_DIR$/network" />
