+++ app/src/main/java/com/example/yandex_school_app/MainScreen.kt
@@ -23,12 +23,12 @@ import com.example.cash_account.presentation.ui.CashAccountScreen
 import com.example.cash_account.presentation.ui.CreateCashAccount
 import com.example.cash_account.presentation.ui.DetailsCashAccountScreen
 import com.example.category.presentation.ui.CategoryScreen
-import com.example.yandex_school_app.features.expense.presentation.ui.DetailsExpenseScreen
-import com.example.yandex_school_app.features.expense.presentation.ui.ExpenseScreen
-import com.example.yandex_school_app.features.expense.presentation.ui.HistoryExpenseScreen
-import com.example.yandex_school_app.features.income.presentation.ui.DetailsIncomeScreen
-import com.example.yandex_school_app.features.income.presentation.ui.HistoryIncomeScreen
-import com.example.yandex_school_app.features.income.presentation.ui.IncomeScreen
+import com.example.expense.presentation.ui.DetailsExpenseScreen
+import com.example.expense.presentation.ui.ExpenseScreen
+import com.example.expense.presentation.ui.HistoryExpenseScreen
+import com.example.income.presentation.ui.DetailsIncomeScreen
+import com.example.income.presentation.ui.HistoryIncomeScreen
+import com.example.income.presentation.ui.IncomeScreen
 import com.example.settings.presentation.SettingsScreen
 import com.example.navigation.BottomNavigationBarCustom
 import com.example.navigation.NavigationCustomItem
