+++ app/src/main/java/com/example/yandex_school_app/di/module/ViewModelModule.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import com.example.yandex_school_app.di.ViewModelKey
 import com.example.cash_account.presentation.AccountViewModel
 import com.example.category.presentation.CategoryViewModel
-import com.example.yandex_school_app.features.expense.presentation.viewmodel.ExpenseViewModel
-import com.example.yandex_school_app.features.income.presentation.viewmodel.IncomeViewModel
+import com.example.expense.presentation.ExpenseViewModel
+import com.example.income.presentation.viewmodel.IncomeViewModel
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
