--- app/src/main/java/com/example/yandex_school_app/features/income/presentation/ui/DetailsIncomeScreen.kt
@@ -1,9 +1,0 @@
-package com.example.yandex_school_app.features.income.presentation.ui
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-fun DetailsIncomeScreen(modifier: Modifier) {
-
-}
