+++ features/expense/build.gradle.kts
@@ -1,5 +1,5 @@
 plugins {
-    alias(libs.plugins.android.application)
+    alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.kotlin.compose)
 
@@ -7,17 +7,14 @@ plugins {
 }
 
 android {
-    namespace = "com.example.yandex_school_app"
+    namespace = "com.example.expense"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.example.yandex_school_app"
         minSdk = 26
-        targetSdk = 35
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
@@ -42,17 +39,10 @@ android {
 }
 
 dependencies {
-    implementation(project(":navigation"))
     implementation(project(":common"))
     implementation(project(":network"))
-    implementation(project(":features:settings"))
-    implementation(project(":features:category"))
-    implementation(project(":features:cash_account"))
-    implementation(project(":features:expense"))
-    implementation(project(":features:income"))
 
     implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.lifecycle.runtime.ktx)
     implementation(libs.androidx.activity.compose)
     implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.ui)
@@ -60,19 +50,10 @@ dependencies {
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
 
-    implementation(libs.airbnb.android.lottie.compose)
-    implementation(libs.androidx.navigation.compose)
     implementation(libs.google.dagger.core)
     ksp(libs.google.dagger.compiler)
-    implementation(libs.squareup.retrofit2)
-    implementation(libs.squareup.retrofit2.gson)
-    implementation(libs.androidx.lifecycle.viewmodel)
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
     androidTestImplementation(libs.androidx.espresso.core)
-    androidTestImplementation(platform(libs.androidx.compose.bom))
-    androidTestImplementation(libs.androidx.ui.test.junit4)
-    debugImplementation(libs.androidx.ui.tooling)
-    debugImplementation(libs.androidx.ui.test.manifest)
 }
