+++ features/expense/src/main/java/com/example/expense/presentation/ui/DetailsExpenseScreen.kt
@@ -1,0 +1,9 @@
+package com.example.expense.presentation.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun DetailsExpenseScreen(modifier: Modifier) {
+
+}
