+++ features/income/src/main/java/com/example/income/presentation/ui/DetailsIncomeScreen.kt
@@ -1,0 +1,9 @@
+package com.example.income.presentation.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+
+@Composable
+fun DetailsIncomeScreen(modifier: Modifier) {
+
+}
