+++ settings.gradle.kts
@@ -28,3 +28,5 @@ include(":features:settings")
 include(":features:category")
 include(":features:cash_account")
 include(":network")
+include(":features:expense")
+include(":features:income")
